HHH-9777 : Copy tests from AbstractDereferencedCollectionTest into subclasses
